Froi/scrolling detail tables (#29)
* Add media queries and static table headers

Added CSS classes to make the table headers static and let the rest of the table scroll
Added media queries to change behavior of frozen table headers son smaller devices like phones. The headers and table will scroll with the rest of the page on phone and smaller devices.

Signed-off-by: Froilan Irizarry Rivera <1918027+froi@users.noreply.github.com>

* Add new CSS classes to detail tables

Added new CSS classes to detail tables to make the table headers fixed position. The rest of the table will now scroll on devices larger than a phone, while the headers stay fixed.

Signed-off-by: Froilan Irizarry Rivera <1918027+froi@users.noreply.github.com>