Fix value formating for date fields
The origin data provider changed the value of the fecha_rd field to be a date formatted one. This broke the way we would format values. I've added an field exception list to avoid numeric formatting of date fields.
Along with the formatting of date fields we are now not showing a graph for date fields. The graph for the fecha_rd field is now not showing in the main and detail sites

Signed-off-by: Froilan Irizarry Rivera <1918027+froi@users.noreply.github.com>